code/week10/requirements.txt
langchain
openai==1.6.1
datasets==2.10.1
pinecone-client==3.1.0
tiktoken==0.5.2
python-dotenv
langchain_text_splitters
chromadb
sentence_transformers
langchainhub